# apis/docs_dist/openapi.json
{"openapi": "3.1.0", "info": {"title": "Meditrack Portal APIs", "description": "3-Tier Backend API for Meditrack utilizing Python/FastAPI", "version": "1.0.0"}, "paths": {"/api/v1/health": {"get": {"summary": "Health Check", "operationId": "health_check_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/patients/": {"get": {"tags": ["Patients"], "summary": "Get All Patients", "operationId": "get_all_patients_api_v1_patients__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["Patients"], "summary": "Create Patient", "operationId": "create_patient_api_v1_patients__post", "requestBody": {"content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Data"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/patients/{patient_id}": {"get": {"tags": ["Patients"], "summary": "Get Patient By Id", "operationId": "get_patient_by_id_api_v1_patients__patient_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Patient Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Patients"], "summary": "Update Patient", "operationId": "update_patient_api_v1_patients__patient_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Patient Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Patients"], "summary": "Delete Patient", "operationId": "delete_patient_api_v1_patients__patient_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "patient_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Patient Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/doctors/": {"get": {"tags": ["Doctors"], "summary": "Get All Doctors", "operationId": "get_all_doctors_api_v1_doctors__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["Doctors"], "summary": "Create Doctor", "operationId": "create_doctor_api_v1_doctors__post", "requestBody": {"content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Data"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/doctors/{doc_id}": {"put": {"tags": ["Doctors"], "summary": "Update Doctor", "operationId": "update_doctor_api_v1_doctors__doc_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Doc Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Doctors"], "summary": "Delete Doctor", "operationId": "delete_doctor_api_v1_doctors__doc_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Doc Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/appointments/": {"get": {"tags": ["Appointments"], "summary": "Get All Appointments", "operationId": "get_all_appointments_api_v1_appointments__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["Appointments"], "summary": "Create Appointment", "operationId": "create_appointment_api_v1_appointments__post", "requestBody": {"content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Data"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/appointments/doctor/{doc_id}": {"get": {"tags": ["Appointments"], "summary": "Get Doctor Appointments", "operationId": "get_doctor_appointments_api_v1_appointments_doctor__doc_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Doc Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/appointments/{appt_id}/status": {"put": {"tags": ["Appointments"], "summary": "Update Appointment Status", "operationId": "update_appointment_status_api_v1_appointments__appt_id__status_put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "appt_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Appt Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/appointments/{appt_id}": {"put": {"tags": ["Appointments"], "summary": "Update Appointment", "operationId": "update_appointment_api_v1_appointments__appt_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "appt_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Appt Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Appointments"], "summary": "Delete Appointment", "operationId": "delete_appointment_api_v1_appointments__appt_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "appt_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Appt Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}, "servers": [{"url": "https://meditrack-portal.loca.lt"}]}